feat: Cambia visualización de registro
